Refactor expression rewriting out of annotator
