+++ src/helpers/annotation/get-inline-notes.ts
@@ -1,0 +1,3 @@
+export const getInlineNotes = (notes: string): string => {
+    return notes.trim().replaceAll('\r', '').replaceAll('\n', '<br>')
+}
