+++ src/helpers/annotation/get-rosetta-type-string.ts
@@ -1,0 +1,18 @@
+export const getRosettaTypeString = (
+    type: string | undefined,
+    optional: boolean | undefined,
+    nullable?: boolean,
+): string => {
+    type ??= 'any'
+    type = type.trim()
+
+    if (nullable) {
+        type += ' | nil'
+    }
+
+    if (optional) {
+        return type.includes('|') ? `(${type})?` : `${type}?`
+    }
+
+    return type
+}
