+++ src/helpers/annotation/is-literal-table.ts
@@ -1,0 +1,9 @@
+import { LuaExpression } from '../../analysis'
+
+export const isLiteralTable = (expr: LuaExpression): boolean => {
+    if (expr.type !== 'literal') {
+        return false
+    }
+
+    return expr.luaType === 'table'
+}
