+++ src/helpers/index.ts
@@ -1,3 +1,4 @@
+export * from './annotation'
 export { arrayToRecord } from './array-to-record'
 export * from './convert-analyzed'
 export * from './convert-rosetta'
